[REFACTOR] LLM 서비스 책임 분리
- Gemini SDK 호출부를 gemini_client 모듈로 분리해 초기화/에러처리 일원화
- 프롬프트 구성 로직을 prompt_builder로 이동해 멀티모달 입력을 모듈화
- llm_service는 orchestration만 담당하도록 단순화
